--- lib/lightning/extensions/default/rate_limiter.ex
@@ -1,11 +1,0 @@
-defmodule Lightning.Extensions.Default.RateLimiter do
-  @moduledoc """
-  Rate limiting stub for Lightning.
-  """
-  @behaviour Lightning.Extensions.RateLimiting
-
-  @impl true
-  def limit_request(_conn, _context, _opts) do
-    :ok
-  end
-end
