+++ lib/lightning_web/channels/run_channel.ex
@@ -5,13 +5,13 @@ defmodule LightningWeb.RunChannel do
   use LightningWeb, :channel
 
   alias Lightning.Credentials
-  alias Lightning.Extensions.RuntimeLimiter
   alias Lightning.Extensions.RuntimeLimiting.Action
   alias Lightning.Extensions.RuntimeLimiting.Context
   alias Lightning.Projects
   alias Lightning.Repo
   alias Lightning.Runs
   alias Lightning.Scrubber
+  alias Lightning.Services.RuntimeLimiter
   alias Lightning.Workers
   alias LightningWeb.RunOptions
   alias LightningWeb.RunWithOptions
